# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tire Management System API",
        "description": "API documentation for the Tire Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/stock/tires": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "List all tires",
                "operationId": "0109e5a43e6886d0bbc4e429a64ee0cd",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tires"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Add a single tire",
                "operationId": "5fdf078ac8177c9c27820e5090bcc7b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "brand",
                                    "model",
                                    "size",
                                    "cost"
                                ],
                                "properties": {
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "serial_number": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "vendor": {
                                        "type": "string"
                                    },
                                    "warehouse_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tire created"
                    }
                }
            }
        },
        "/api/stock/tires/{id}/label": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get Tire QR Code Label",
                "operationId": "bbf4d4de21b6103562d88570b1ed1105",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "QR Code Image",
                        "content": {
                            "image/png": {}
                        }
                    }
                }
            }
        },
...
        "/api/stock/tires/{id}/history": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get Tire Movement History",
                "operationId": "761996491c93934e2e601e2f950ef01d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "History list"
                    }
                }
            }
        },
        "/api/stock/tires/{id}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get single tire details",
                "operationId": "c0e1a7088bd8a6af5ec1cf6497120f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tire details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stock"
                ],
                "summary": "Update tire details",
                "operationId": "7166808fe98e0f32ce28be653258be4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "brand": {
                                        "type": "string"
                                    },
                                    "model": {
                                        "type": "string"
                                    },
                                    "size": {
                                        "type": "string"
                                    },
                                    "cost": {
                                        "type": "number"
                                    },
                                    "vendor": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tire updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stock"
                ],
                "summary": "Delete a tire",
                "operationId": "2e23389788fb4519a76027eede6c46f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tire deleted"
                    }
                }
            }
        },
        "/api/stock/tires/bulk": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Bulk Upload Tires (CSV)",
                "operationId": "072f1e44b24345c3382169989c27d38e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bulk created"
                    }
                }
            }
        },
        "/api/stock/tires/{id}/lifecycle": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get Tire Lifecycle (Operations)",
                "operationId": "59b7edbb28a511c06940ce836d02fcd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lifecycle events"
                    }
                }
            }
        },
        "/api/stock/tires/{id}/operations": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get all operations for a tire",
                "operationId": "c8bb5f8b1e4912d9e44b2d443bf1e2ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of operations"
                    }
                }
            }
        },
        "/api/tires/{id}/inspections": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get all inspections for a tire",
                "operationId": "55cce22c0176baae441b8156f209fd58",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of inspections"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}